requirements.txt
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
awscli==1.38.17
blinker==1.9.0
boto3==1.37.17
botocore==1.37.17
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.8
dnspython==2.7.0
docutils==0.16
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.11
filelock==3.16.1
flasgger==0.9.7.1
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-MySQLdb==2.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.8
MarkupSafe==3.0.2
mistune==3.1.2
motor==3.7.0
mysqlclient==2.2.7
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
passlib==1.7.4
platformdirs==4.3.6
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
PyMySQL==1.1.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.7.2
s3transfer==0.11.4
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.6
Werkzeug==3.1.3
